cdコマンドで使う関数を役割ごとにファイルを分けた
