+++ builtin.h
@@ -16,6 +16,9 @@ bool			is_builtin_command(char *cmd_name);
 t_builtin_cmd	*get_builtin_func(char *cmd_name);
 int				builtin_echo(char **argv);
 int				builtin_cd(char **argv);
+bool			change_directory(char *dest);
+bool			will_search_cdpath(char **argv, char *dest);
+int				cd_cdpath_env(char *dest_path);
 int				builtin_pwd(char **argv);
 int				builtin_env(char **argv);
 int				builtin_export(char **argv);
