+++ builtin_cd.c
@@ -20,114 +20,6 @@ static void	put_cd_errmsg(char *dest_path)
 	free(errmsg);
 }
 
-/* get_cd_dest() で取得した移動先パス(絶対or相対) を元に
- * 移動先の絶対パスを返す.
- *
- * dest: get_cd_dest() で取得した移動先.
- * is_canon_path: 正規化されたパスかどうか
- *
- * Return: 移動先ディレクトリの絶対パス.
- */
-static char	*get_cd_abs_dest(char *dest, bool *is_canon_path)
-{
-	char	*physical_path;
-	char	*canon_path;
-
-	if (dest[0] =='/')
-		physical_path = ft_strdup(dest);
-	else
-		physical_path = path_join(g_cwd, dest);
-	if (!physical_path)
-		return (NULL);
-	canon_path = canonicalize_path(physical_path);
-	if (canon_path)
-	{
-		free(physical_path);
-		*is_canon_path = true;
-		return (canon_path);
-	}
-	free(canon_path);
-	*is_canon_path = false;
-	return (physical_path);
-}
-
-static int	set_new_pwd(char *path, bool is_canon_path, bool is_abs_path)
-{
-	char	*new_cwd;
-
-	new_cwd = NULL;
-	if (is_abs_path)
-	{
-		if (is_canon_path == false)
-			new_cwd = get_cwd_path("cd");
-		if (is_canon_path || new_cwd == NULL)
-			new_cwd = ft_strdup(path);
-	}
-	else
-	{
-		new_cwd = get_cwd_path("cd");
-		if (!new_cwd)
-			new_cwd = ft_strdup(path);
-	}
-	if (!new_cwd)
-		return (ERROR);
-	set_cwd(new_cwd);
-	free(new_cwd);
-	return (SUCCESS);
-}
-
-/* 実際に chdir() を実行して現在のプロセスのcwdを変更する
- *
- * abs_dest: chdir() を実行する絶対パス.
- * arg: chdir(abs_dest) が失敗した時に移動するパス.
- *   これは bash の cd がabs_dest に移動失敗した時に
- *   引数で chdir() を試す挙動に合わせている.
- * is_canon_path: 正規化されたパスかどうか
- *
- * Return: 移動に成功したかどうか.
- */
-static bool	change_dir_process(char *abs_dest,
-	const char *arg, bool is_canon_path)
-{
-	int	status;
-	int	old_errno;
-
-	status = chdir(abs_dest);
-	if (status == 0)
-	{
-		set_new_pwd(abs_dest, is_canon_path, true);
-		return (true);
-	}
-	old_errno = errno;
-	status = chdir(arg);
-	if (status == 0)
-	{
-		set_new_pwd(abs_dest, is_canon_path, false);
-		return (true);
-	}
-	errno = old_errno;
-	return (false);
-}
-
-/* プロセスのcwdを変更する
- *
- * dest: get_cd_dest() で取得した移動先.
- *
- * Return: 移動に成功したかどうか.
- */
-static bool	change_directory(char *dest)
-{
-	char	*path;
-	int		status;
-	bool	is_canon_path;
-
-	path = get_cd_abs_dest(dest, &is_canon_path);
-	PRINT_DEBUG("path: |%s|, is_canon_path: %d\n", path, is_canon_path);
-	status = change_dir_process(path, dest, is_canon_path);
-	free(path);
-	return (status);
-}
-
 /* cd先のディレクトリをcdコマンドのargvを元に作成して返す.
  *
  * argv: builtin_cd() の引数.
@@ -149,69 +41,6 @@ static char	*get_cd_dest_from_argv(char **argv)
 		return (dest_path);
 	}
 	return (ft_strdup(argv[1]));
-}
-
-/* $CDPATH を検索するかどうかを返す
- *
- * argv: builtin_cd() の引数.
- * dest: get_cd_dest() で取得した移動先パス.
- *
- * Return: $CDPATH を検索する必要があるかどうか.
- */
-static bool	will_search_cdpath(char **argv, char *dest)
-{
-	if (argv[1] == NULL || argv[1][0] == '/')
-		return (false);
-	if (ft_strcmp((char *)dest, ".") == 0 ||
-		ft_strcmp((char *)dest, "..") == 0 ||
-		ft_strncmp((char *)dest, "./", 2) == 0 ||
-		ft_strncmp((char *)dest, "../", 3) == 0)
-		return (false);
-	return (true);
-}
-
-/* $CDPATH を元に移動する.
- *
- * Return: 移動に成功したら true を返す.
- *   そうでない場合は false を返す.
- */
-static int	cd_cdpath_env(char *dest_path)
-{
-	char	*cdpath;
-	int		i;
-	char	*abs_path;
-	char	**dirs;
-	char	*tmp;
-
-	cdpath = get_env_val("CDPATH");
-	if (!cdpath)
-		return (false);
-	dirs = get_colon_units(cdpath);
-	free(cdpath);
-	i = 0;
-	while (dirs[i])
-	{
-		if (dirs[i][0] != '/')
-		{
-			tmp = path_join(g_cwd, dirs[i]);
-			abs_path = path_join(tmp, dest_path);
-			free(tmp);
-		}
-		else
-			abs_path = path_join(dirs[i], dest_path);
-		if (change_directory(abs_path))
-		{
-			if (ft_strlen(dirs[i]))
-				printf("%s\n", g_cwd);
-			free(abs_path);
-			free_ptrarr((void **)dirs);
-			return (true);
-		}
-		free(abs_path);
-		i++;
-	}
-	free_ptrarr((void **)dirs);
-	return (false);
 }
 
 int	builtin_cd(char **argv)
